+++ tsup.config.ts
@@ -9,6 +9,7 @@ export default defineConfig({
         checkout: 'src/checkout.ts',
         helpers: 'src/helpers.ts',
         license: 'src/license.ts',
+        oauth: 'src/oauth.ts',
         react: 'src/react.ts',
         vue: 'src/vue.ts',
     },